Move api.ts to api/remote-ac.ts
